Add Tracing Utilities
The patch adds tracing utilities, which allows us to analyze receipts with full syntactic analysis matched to op codes. For example, if the code looked like this:

```
return balanceOf[msg.sender];
```

This might look like the following opcodes:

```
CALLER  -- load msg.sender to the stack
MSTORE  -- push msg.sender to memory
SHA3    -- run sha3 of msg.sender
SLOAD   -- load from that address
```

The trace code can analyze this and realize that the SLOAD matched with that line of code looks like:

`SLOAD(SHA3(sender))`

The goal of the tracing utilities are to bring this analysis into your test cases to verify the code works as expected, thus you could run a trace:

```javascript
saddle.trace(getBalanceReceipt, {
  preFilter: ({op}) => op === 'SLOAD', // ignore any op code that isn't a storage load
  postFilter: (log) => log.source.includes('balanceOf[msg.sender]'), // filter to storage loads that match this line of code
  exec: (logs) => {
    expect(logs.length).toEqual(1); // check that there is exactly one storage load associated with the line of code
  }
});
```

But that's only the beginning. There's a ton more than can be done and built. Hopefully tracing can be a big part of answering "I think this is what Solidity does, but I'm not sure"-type questions.