Fix test trace, bump to 0.1.10
This patch adds a variety of fixes to our tracing code to add better support for calls and delegatecalls. We keep a stack of the pc->source code mappings each time we jump forward in a call or delegate call. I'm not entirely sure how sol-coverage handles this, but this approach seems to work fairly robustly on tests of the tracing system